requirements.txt
selenium
webdriver-manager
beautifulsoup4
pandas
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client